backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlmodel==0.0.24
apscheduler==3.11.0
httpx==0.28.1
python-dotenv==1.1.0
python-dateutil==2.9.0
